Add files via upload
We have do some changes in order to run the output video we prefer to use VS Code for the execution process. In that we have to update the Frame and Input video path from the local device and the all other details are pre install but in some cases you  have to istall numy npm and other modules in the terminal which can be get from online resorces and generative ai by using it you can clear all your error to run the project.